+++ leo-picture-backend/src/main/java/com/leo/leopicturebackend/service/AiService.java
@@ -12,10 +12,10 @@ public interface AiService {
     /**
      * 处理AI聊天流请求
      *
-     * @param memoryId 内存ID
+     * @param memoryId 内存ID（使用long类型以支持大数值）
      * @param userMessage 用户消息
      * @param request HTTP请求
      * @return ServerSentEvent流
      */
-    Flux<ServerSentEvent<String>> chatStream(int memoryId, String userMessage, HttpServletRequest request);
+    Flux<ServerSentEvent<String>> chatStream(long memoryId, String userMessage, HttpServletRequest request);
 }
